Struggling to seed with correct www
